# Code/gamesUP/docs/openapi.yaml
openapi: 3.0.1
info:
  title: GamesUP API
  description: |
    API REST du catalogue et des parcours clients GamesUP.
    Les ressources personnelles sont toujours déterminées par le jeton JWT :
    aucune route ne permet de sélectionner un autre utilisateur.
  contact:
    name: GamesUP
  license:
    name: Projet pédagogique
  version: 1.0.0
servers:
- url: /
  description: Serveur courant
...
paths:
  /api/v1/users/me/wishlist/{gameId}:
    put:
      tags:
      - Wishlist
      summary: Ajouter un jeu
      description: "Opération idempotente : un ajout répété ne crée pas de doublon."
      operationId: addGame
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Wishlist
      summary: Retirer un jeu
      description: La suppression d'un jeu absent reste idempotente.
      operationId: removeGame
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/reviews/{reviewId}:
    put:
      tags:
      - Owned reviews
      summary: Modifier son avis
      operationId: update
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Owned reviews
      summary: Supprimer son avis
      operationId: delete
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/games/{gameId}/reviews:
    get:
      tags:
      - Game reviews
      summary: Lister les avis publiés
      description: Les avis masqués sont exclus et seul le prénom public du client
        est exposé.
      operationId: findPublishedByGame
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseReviewResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
      - Game reviews
      summary: Créer son avis
      description: Un seul avis est autorisé par couple client/jeu.
      operationId: create_1
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/users/me:
    get:
      tags:
      - Current user
      summary: Consulter son profil
      operationId: findCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Current user
      summary: Modifier partiellement son profil
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/users/me/wishlist:
    get:
      tags:
      - Wishlist
      summary: Consulter sa wishlist
      operationId: findCurrentWishlist
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
components:
  schemas:
    GameSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
        minPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        minAge:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
        publisher:
          type: string
        authors:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
    WishlistResponse:
      type: object
      properties:
        games:
          type: array
          items:
            $ref: "#/components/schemas/GameSummaryResponse"
    ReviewRequest:
      required:
      - rating
      type: object
      properties:
        rating:
          maximum: 5
          exclusiveMaximum: false
          minimum: 1
          exclusiveMinimum: false
          type: number
          description: Note entière de 1 à 5.
          example: 5
        comment:
          maxLength: 2000
          minLength: 0
          type: string
          example: "Accessible, rapide et très rejouable."
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        gameId:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
        comment:
          type: string
        reviewerName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserProfileUpdateRequest:
      type: object
      properties:
        email:
          maxLength: 320
          minLength: 0
          type: string
          example: alice.martin@example.com
        firstName:
          maxLength: 100
          minLength: 0
          type: string
          example: Alice
        lastName:
          maxLength: 100
          minLength: 0
          type: string
          example: Martin
        validPatch:
          type: boolean
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum:
          - CLIENT
          - ADMIN
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    PageResponseReviewResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Jeton retourné par l'inscription ou la connexion.
      scheme: bearer
      bearerFormat: JWT